+++ src/components/navbar/Navbar.tsx
@@ -13,7 +13,10 @@ import { MobileMenu } from "./MobileMenu";
 import { modalNames, modalStore } from "../../store/modalStore";
 import ConnectedWallets from "./ConnectedWallets";
 import { useStarknetWallet } from "../../hooks/useStarknetWallet";
-import { useBitcoinWallet } from "@gardenfi/wallet-connectors";
+import {
+  useBitcoinWallet,
+  useLitecoinWallet,
+} from "@gardenfi/wallet-connectors";
 import { useSolanaWallet } from "../../hooks/useSolanaWallet";
 import { viewPortStore } from "../../store/viewPortStore";
 import { useSuiWallet } from "../../hooks/useSuiWallet";
@@ -21,6 +24,7 @@ import { useTronWallet } from "../../hooks/useTronWallet";
 
 export const Navbar = () => {
   const { isConnected, address } = useEVMWallet();
+  const { account: ltcAddress } = useLitecoinWallet();
   const { starknetAddress } = useStarknetWallet();
   const { account: btcAddress } = useBitcoinWallet();
   const { solanaAddress } = useSolanaWallet();
@@ -86,6 +90,7 @@ export const Navbar = () => {
         {address ||
         starknetAddress ||
         btcAddress ||
+        ltcAddress ||
         solanaAddress ||
         suiConnected ||
         tronConnected ? (