+++ src/constants/api.ts
@@ -1,11 +1,8 @@
 import { Network, Url } from "@gardenfi/utils";
 
 const REQUIRED_ENV_VARS = {
-  STAKING_URL: import.meta.env.VITE_STAKING_URL,
-  INFO_URL: import.meta.env.VITE_INFO_URL,
-  QUOTE_URL: import.meta.env.VITE_QUOTE_URL,
   BASE_URL: import.meta.env.VITE_BASE_URL,
-  REWARD: import.meta.env.VITE_REWARD_URL,
+  QUOTE_URL: import.meta.env.VITE_QUOTE_URL,
   EXPLORER: import.meta.env.VITE_EXPLORER_URL,
   API_KEY: import.meta.env.VITE_API_KEY,
   BALANCE_URL: import.meta.env.VITE_BALANCE_URL,
@@ -19,81 +16,68 @@ export const API = () => {
   return {
     api_key: REQUIRED_ENV_VARS.API_KEY,
     home: new Url("https://garden.finance"),
-    data: {
-      chains: () => new Url(REQUIRED_ENV_VARS.BASE_URL).endpoint("/v2/chains"),
-      blockNumbers: (network: "mainnet" | "testnet" | "localnet") =>
-        new Url(REQUIRED_ENV_VARS.INFO_URL)
-          .endpoint("blocknumbers")
-          .endpoint(network),
-      notification: () =>
-        new Url(REQUIRED_ENV_VARS.INFO_URL).endpoint("notification"),
-    },
+    data: (() => {
+      const infoBase = () =>
+        new Url(REQUIRED_ENV_VARS.BASE_URL).endpoint("info");
+      return {
+        chains: () =>
+          new Url(REQUIRED_ENV_VARS.BASE_URL).endpoint("/v2/chains"),
+        blockNumbers: (network: "mainnet" | "testnet" | "localnet") =>
+          infoBase().endpoint("blocknumbers").endpoint(network),
+        notification: () => infoBase().endpoint("notification"),
+      };
+    })(),
     buildId: "/build-id.json",
     baseUrl: new Url(REQUIRED_ENV_VARS.BASE_URL),
     quote: {
-      quote: new Url(REQUIRED_ENV_VARS.BASE_URL),
-      fiatValues: new Url(REQUIRED_ENV_VARS.QUOTE_URL).endpoint("/fiat"),
+      quote: new Url(REQUIRED_ENV_VARS.BASE_URL).endpoint("quote"),
+      fiatValues: new Url(REQUIRED_ENV_VARS.QUOTE_URL).endpoint("fiat"),
     },
-    balance: {
-      bitcoin: (address: string) =>
-        new Url(REQUIRED_ENV_VARS.BALANCE_URL)
-          .endpoint("balances")
-          .endpoint("bitcoin")
-          .endpoint(address),
-      evm: (address: string) =>
-        new Url(REQUIRED_ENV_VARS.BALANCE_URL)
-          .endpoint("balances")
-          .endpoint("evm")
-          .endpoint(address),
-      starknet: (address: string) =>
-        new Url(REQUIRED_ENV_VARS.BALANCE_URL)
-          .endpoint("balances")
-          .endpoint("starknet")
-          .endpoint(address),
-      solana: (address: string) =>
-        new Url(REQUIRED_ENV_VARS.BALANCE_URL)
-          .endpoint("balances")
-          .endpoint("solana")
-          .endpoint(address),
-      sui: (address: string) =>
-        new Url(REQUIRED_ENV_VARS.BALANCE_URL)
-          .endpoint("balances")
-          .endpoint("sui")
-          .endpoint(address),
-      tron: (address: string) =>
-        new Url(REQUIRED_ENV_VARS.BALANCE_URL)
-          .endpoint("balances")
-          .endpoint("tron")
-          .endpoint(address),
-    },
-    stake: {
-      stakePosition: (userId: string) =>
-        new Url("stake", REQUIRED_ENV_VARS.STAKING_URL)
-          .endpoint("stakes")
-          .addSearchParams({
-            userId: userId.toLowerCase(),
-          }),
-      globalApy: new Url(REQUIRED_ENV_VARS.STAKING_URL).endpoint("apy"),
-      stakeApy: (address: string) =>
-        new Url(REQUIRED_ENV_VARS.STAKING_URL)
-          .endpoint("apy")
-          .endpoint(address.toLowerCase()),
-      stakingStats: new Url(REQUIRED_ENV_VARS.STAKING_URL)
-        .endpoint("stake")
-        .endpoint("staking-stats"),
-      accumulatedReward: (userId: string) =>
-        new Url(REQUIRED_ENV_VARS.STAKING_URL)
-          .endpoint("rewards")
-          .endpoint("accumulated")
-          .endpoint(userId),
-      reward: (userId: string) =>
-        new Url(REQUIRED_ENV_VARS.STAKING_URL)
-          .endpoint("rewards")
-          .endpoint(userId),
-      epoch: new Url(REQUIRED_ENV_VARS.STAKING_URL)
-        .endpoint("rewards")
-        .endpoint("epochs"),
-    },
+    balance: (() => {
+      const balancesBaseApi = () =>
+        new Url(REQUIRED_ENV_VARS.BALANCE_URL).endpoint("balances");
+      return {
+        bitcoin: (address: string) =>
+          balancesBaseApi().endpoint("bitcoin").endpoint(address),
+        evm: (address: string) =>
+          balancesBaseApi().endpoint("evm").endpoint(address),
+        starknet: (address: string) =>
+          balancesBaseApi().endpoint("starknet").endpoint(address),
+        solana: (address: string) =>
+          balancesBaseApi().endpoint("solana").endpoint(address),
+        sui: (address: string) =>
+          balancesBaseApi().endpoint("sui").endpoint(address),
+        tron: (address: string) =>
+          balancesBaseApi().endpoint("tron").endpoint(address),
+      };
+    })(),
+    stake: (() => {
+      const distributorBase = () =>
+        new Url(REQUIRED_ENV_VARS.BASE_URL).endpoint("distributor");
+      return {
+        stakePosition: (userId: string) =>
+          distributorBase()
+            .endpoint("stake")
+            .endpoint("stakes")
+            .addSearchParams({
+              userId: userId.toLowerCase(),
+            }),
+        globalApy: distributorBase().endpoint("apy"),
+        stakeApy: (address: string) =>
+          distributorBase().endpoint("apy").endpoint(address.toLowerCase()),
+        stakingStats: distributorBase()
+          .endpoint("stake")
+          .endpoint("staking-stats"),
+        accumulatedReward: (userId: string) =>
+          distributorBase()
+            .endpoint("rewards")
+            .endpoint("accumulated")
+            .endpoint(userId),
+        reward: (userId: string) =>
+          distributorBase().endpoint("rewards").endpoint(userId),
+        epoch: distributorBase().endpoint("rewards").endpoint("epochs"),
+      };
+    })(),
     explorer: (orderId: string) =>
       new Url("order", REQUIRED_ENV_VARS.EXPLORER).endpoint(orderId),
     mempoolTxExplorer: (network: Network, txHash: string) =>